move types/ReviewProvider content to utils/reviewProvider
